requirements.txt
diffusers==0.18.1
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
matplotlib==3.5.1
pillow==9.3.0
scikit-learn==1.3.0
ipykernel==6.29.4
notebook==7.1.2
opencv-contrib-python==4.8.1.78
ipywidgets==8.1.2